+++ lib/td_dd/data_structures/data_structures.ex
@@ -1127,6 +1127,12 @@ defmodule TdDd.DataStructures do
   defp add_filter({"updated_at", updated_at}, query),
     do: where(query, [sn], sn.updated_at >= ^updated_at)
 
+  defp add_filter({"system_id", system_id}, query) do
+    query
+    |> join(:inner, [sn], ds in assoc(sn, :data_structure))
+    |> where([_sn, ds], ds.system_id == ^system_id)
+  end
+
   defp add_filter(_, query), do: query
 
   @doc """
@@ -1211,6 +1217,7 @@ defmodule TdDd.DataStructures do
   """
 
   def bulk_update_structure_note(%StructureNote{} = structure_note, attrs, user_id) do
+    structure_note = Repo.preload(structure_note, :data_structure)
     changeset = StructureNote.bulk_update_changeset(structure_note, attrs)
 
     if changeset.changes == %{} do
@@ -1282,6 +1289,7 @@ defmodule TdDd.DataStructures do
   end
 
   def update_structure_note(%StructureNote{} = structure_note, attrs, user_id) do
+    structure_note = Repo.preload(structure_note, :data_structure)
     changeset = StructureNote.changeset(structure_note, attrs)
 
     Multi.new()