+++ lib/td_dd/grants.ex
@@ -19,6 +19,7 @@ defmodule TdDd.Grants do
   alias TdDd.Grants.GrantRequestGroup
   alias TdDd.Grants.GrantRequestStatus
   alias TdDd.Repo
+  alias TdDd.Search.IndexWorker
 
   def get_grant!(id, opts \\ []) do
     Grant
@@ -39,6 +40,19 @@ defmodule TdDd.Grants do
     |> Multi.insert(:grant, changeset)
     |> Multi.run(:audit, Audit, :grant_created, [user_id])
     |> Repo.transaction()
+    |> reindex_grants()
+  end
+
+  defp reindex_grants({:ok, %{grant: %Grant{id: id}} = multi}) do
+    IndexWorker.reindex_grants(id)
+    {:ok, multi}
+  end
+
+  defp reindex_grants(error), do: error
+
+  defp on_delete({:ok, %{grant: %Grant{id: id}} = multi}) do
+    IndexWorker.delete_grants(id)
+    {:ok, multi}
   end
 
   def update_grant(%Grant{} = grant, params, %Claims{user_id: user_id}) do
@@ -48,6 +62,7 @@ defmodule TdDd.Grants do
     |> Multi.update(:grant, changeset)
     |> Multi.run(:audit, Audit, :grant_updated, [changeset, user_id])
     |> Repo.transaction()
+    |> reindex_grants()
   end
 
   def delete_grant(%Grant{data_structure: data_structure} = grant, %Claims{user_id: user_id}) do
@@ -58,6 +73,7 @@ defmodule TdDd.Grants do
     |> Multi.delete(:grant, grant)
     |> Multi.run(:audit, Audit, :grant_deleted, [user_id])
     |> Repo.transaction()
+    |> on_delete
   end
 
   def list_grant_request_groups do