+++ lib/td_dd_web/router.ex
@@ -100,9 +100,14 @@ defmodule TdDdWeb.Router do
     end
 
     get("/data_structures/search/reindex_all", SearchController, :reindex_all)
+    get("/grants/search/reindex_all", SearchController, :reindex_all_grants)
 
     get("/data_structure_filters", DataStructureFilterController, :index)
     post("/data_structure_filters/search", DataStructureFilterController, :search)
+
+    get("/grant_filters", GrantFilterController, :index)
+    post("/grant_filters/search", GrantFilterController, :search)
+    post("/grant_filters/search/mine", GrantFilterController, :search_mine)
 
     get("/data_structure_user_filters/user/me", UserSearchFilterController, :index_by_user)
     resources("/data_structure_user_filters", UserSearchFilterController, except: [:new, :edit])
@@ -113,7 +118,9 @@ defmodule TdDdWeb.Router do
 
     resources("/data_structure_tags", DataStructureTagController, except: [:new, :edit])
 
-    resources("/grants", GrantController, except: [:create, :new, :edit, :index])
+    resources("/grants", GrantController, except: [:create, :new, :edit])
+    post("/grants/search", SearchController, :search_grants)
+    post("/grants/search/mine", SearchController, :search_my_grants)
 
     resources("/grant_request_groups", GrantRequestGroupController,
       only: [:index, :show, :create, :delete]